pyproject.toml
[project]
name = "marianne-ai-compose"
version = "0.1.0"
description = "Orchestration tool for Claude AI sessions"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.11"
authors = [
    { name = "Emzi" }
]
keywords = ["claude", "ai", "orchestration", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",

    # CLI
    "typer[all]>=0.9.0",
    "rich>=13.0.0",

    # HTTP (for API backend + notifications)
    "httpx>=0.27.0",
    "anthropic>=0.87.0",         # CVE-2026-34450, CVE-2026-34452

    # Dashboard
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",

    # Desktop notifications
    "plyer>=2.1.0",

    # Structured logging
    "structlog>=24.1.0",

    # Security minimum versions — transitive deps with known CVEs (F-061, F-137)
    "cryptography>=46.0.6",   # CVE-2026-34073 (F-061)
    "pyjwt>=2.12.0",          # CVE-2026-32597 (F-061)
    "requests>=2.33.0",       # CVE-2026-25645 (F-061)
    "pygments>=2.20.0",       # CVE-2026-4539 (F-137)
]

[project.optional-dependencies]
daemon = ["psutil>=5.9", "watchfiles>=0.21"]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
    "mkdocs-gen-files>=0.5",
    "mkdocs-literate-nav>=0.6",
    "pymdown-extensions>=10.21.2",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.3.0",
    "hypothesis>=6.100.0",
    "pytest-xdist>=3.5.0",
    "pytest-randomly>=3.15.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
mzt = "marianne.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/marianne"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = [
    "B008",  # Function call in default argument - required for typer CLI patterns
    "E402",  # Module-level import not at top - logger init before imports is intentional
    "N802",  # Function name should be lowercase - we override stdlib methods (doRollover)
    "N803",  # Argument name should be lowercase - we match stdlib signatures (maxBytes, backupCount)
    "N806",  # Variable in function should be lowercase - Pydantic model names etc.
    "N817",  # CamelCase imported as acronym - domain convention
    "N818",  # Exception name should end in Error - legacy names
    "SIM102",  # Nested if - sometimes clearer with intermediate comments
    "SIM105",  # contextlib.suppress - try/except/pass is clearer for error recovery
    "SIM108",  # Ternary operator - multi-line if/else often more readable
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
timeout = 30  # Safety net: kill any test that runs >30s (prevents orphaned processes)
addopts = "--tb=short -n auto -m 'not playwright'"
# Coverage: run explicitly with `pytest --cov=mzt --cov-report=term-missing:skip-covered --cov-report=json:coverage.json --no-cov-on-fail`
# Playwright tests: run separately with `pytest -m playwright -n0`
markers = [
    "adversarial: tests with adversarial/garbage inputs",
    "smoke: basic smoke tests for critical paths",
    "overnight: long-running tests for cron execution",
    "property_based: hypothesis property-based tests",
    "playwright: browser E2E tests — must run with -n0 (no xdist)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::RuntimeWarning",
    "ignore::UserWarning:pydantic.*:",
    "ignore::UserWarning:rich.*:",
    "ignore::pytest.PytestWarning",
    "ignore::FutureWarning",
    "ignore::UnicodeWarning",
]

[tool.coverage.run]
source = ["src/marianne"]
omit = [
    "src/marianne/dashboard/static/*",
    "src/marianne/dashboard/templates/*",
    "src/marianne/scores/templates/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "@overload",
    "\\.\\.\\.",
]
show_missing = true
skip_covered = true
fail_under = 80
